+++ NEWS.md
@@ -11,6 +11,8 @@
 * `new_scale()` does not fail when using the default mapping (#7, thanks @
 RomanAbashin)
 
+* `new_scale()` adds a default scale if not supplied
+
 # ggnewscale 0.2.0 - C major
 
 ## Bugfixes